requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
discord-webhook==1.3.0
aiohttp==3.8.5
beautifulsoup4==4.12.2
python-telegram-bot==20.6
apscheduler==3.10.4
aiosqlite==0.19.0
xgboost==2.0.0
lightgbm==4.1.0
catboost==1.2.2
plotly==5.17.0
scipy==1.11.3
kaleido==0.2.1
optuna==3.3.0
shap==0.42.1